requirements.txt
Flask
flask_sqlalchemy
openai
python-dotenv